+++ context/CONTEXT_1771068993.md
@@ -1,0 +1,10 @@
+# Decompose Agent Context
+
+## Target File
+`js/ui/components/VideoModal.js` (approx 6334 lines).
+
+## Rationale
+This is the second largest file in the codebase. The largest (`profileModalController.js`) was recently refactored with a facade pattern. `VideoModal.js` was touched recently but likely only for minor updates or formatting, and contains distinct blocks of logic suitable for extraction.
+
+## Plan
+Extract "Similar Video" related logic into a helper module.
