+++ decisions/DECISIONS_1771068993.md
@@ -1,0 +1,23 @@
+# Decisions Log - Decompose Agent
+
+## Changes
+Extracted "Similar Video" and pointer derivation logic from \`js/ui/components/VideoModal.js\` into a new module \`js/ui/components/VideoModalSimilarHelpers.js\`.
+
+## Rationale
+- \`js/ui/components/VideoModal.js\` was the second largest file (6334 lines).
+- Similar Video logic formed cohesive blocks with clear dependencies.
+- Pointer derivation was a helper function reused internally.
+
+## Implementation Details
+- Created \`js/ui/components/VideoModalSimilarHelpers.js\` containing:
+  - \`derivePointerKeyFromInput\`
+  - \`buildSimilarCardIdentity\`
+  - \`prepareSimilarVideoCard\`
+  - \`attachSimilarCardViewCounter\`
+  - \`getViewCountLabel\`
+  - \`formatViewCountLabel\`
+- Updated \`js/ui/components/VideoModal.js\` to import these functions and pass dependencies (\`this.helpers\`, \`this.DEFAULT_PROFILE_AVATAR\`, \`this.dispatch\`, \`this.log\`).
+- Updated \`scripts/check-file-size.mjs\` baseline for \`VideoModal.js\` to 6045 lines.
+
+## Next Steps
+- Further decomposition of \`VideoModal.js\` (e.g. More Menu logic) in future runs.
