+++ docs/agents/task-logs/daily/2026-02-14_11-46-04_decompose-agent_completed.md
@@ -1,0 +1,24 @@
+# Decompose Agent Run
+
+**Status**: Completed
+**Agent**: decompose-agent
+**Date**: 2026-02-14
+
+## Summary
+Decomposed `js/ui/components/VideoModal.js` by extracting "Similar Video" and "Pointer Key" logic into a new helper module `js/ui/components/VideoModalModalSimilarHelpers.js`.
+
+## Outcome
+- Reduced `js/ui/components/VideoModal.js` line count from 6334 to 6045 lines (-289 lines).
+- Created `tests/ui/components/VideoModalSimilarHelpers.test.mjs` with comprehensive unit tests for the extracted logic.
+- Updated baseline in `scripts/check-file-size.mjs`.
+
+## Verification
+- `npm run lint`: Passed.
+- `npm run test:unit`: Passed (including new tests).
+- `scripts/check-file-size.mjs --report`: Verified reduction.
+
+## Artifacts
+- `test_logs/TEST_LOG_1771068993.md`
+- `decisions/DECISIONS_1771068993.md`
+- `context/CONTEXT_1771068993.md`
+- `todo/TODO_1771068993.md`
