+++ test_logs/TEST_LOG_1771068993.md
@@ -1,0 +1,22 @@
+# Test Log - Decompose Agent
+
+## Summary
+Decomposition of \`js/ui/components/VideoModal.js\` into \`js/ui/components/VideoModalSimilarHelpers.js\`.
+
+## Verification Steps
+
+### 1. Lint
+Command: \`npm run lint\`
+Result: Passed.
+
+### 2. Unit Tests
+Command: \`npm run test:unit\`
+Result: Passed.
+- Existing tests passed.
+- New test \`tests/ui/components/VideoModalSimilarHelpers.test.mjs\` passed (17 subtests).
+
+### 3. File Size Check
+Command: \`node scripts/check-file-size.mjs --report\`
+Result: Passed.
+- \`js/ui/components/VideoModal.js\` reduced from 6334 lines to 6045 lines (-289 lines).
+- Baseline in \`scripts/check-file-size.mjs\` updated to 6045.
