+++ todo/TODO_1771068993.md
@@ -1,0 +1,28 @@
+# Decompose Agent Todo
+
+## Extraction: Similar Video Helpers
+Target file: `js/ui/components/VideoModalSimilarHelpers.js`
+
+### Functions to Extract
+1. `buildSimilarCardIdentity(video, overrides, { helpers, defaultAvatar })`
+   - Originally `buildSimilarCardIdentity`
+   - Depends on `this.helpers` and `this.DEFAULT_PROFILE_AVATAR`.
+   - Action: Refactor to accept these as arguments.
+
+2. `derivePointerKeyFromInput(pointer)`
+   - Used by similar video logic (and others?).
+   - Seems pure (no `this` dependency except maybe internal calls?).
+   - Checked code: `derivePointerKeyFromInput` calls `this.derivePointerKeyFromInput` recursively. It is pure.
+
+3. `prepareSimilarVideoCard(card, meta, index, { dispatchCallback })`
+   - Originally `prepareSimilarVideoCard`.
+   - Uses `this.dispatch`.
+   - Action: Pass a dispatch callback.
+
+4. `attachSimilarCardViewCounter(card, pointerInfo, { subscribeToVideoViewCount, formatViewCount, logger })`
+   - Originally `attachSimilarCardViewCounter`.
+   - Uses `subscribeToVideoViewCount` (global import?), `this.getViewCountLabel`, `this.formatViewCountLabel`, `this.log`.
+   - Action: Extract view count formatting logic too, or pass dependencies.
+
+## Goal
+Reduce `VideoModal.js` size by extracting these substantial methods.
